tcpmv2: pe: Add mux_set call following successful DATA_RESET
This CL adds a call to set_usb_mux_with_current_data_role after
performing a successful DATA_RESET. This is being added to satisfy
6.3.14 Data_Reset Message, item #3:
If the Port was operating in [USB 3.2] or [USB4] reapply the port’s
Rx Terminations (see [USB 3.2]).

Without this call a mode entry attempt that is either not successful,
or rejected by the port partner would leave the USB3.x SS lanes
disconnected. This would also be the case following an Exit_Mode
request from the AP when DATA_RESET is supported.

BUG=b:194504052
BRANCH=none
TEST=manual

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I3b6f9d00beb5fcc5a68c44887bde03d22b61d3e6