zephyr: herobrine: Handle HPD High embedded in DP-Status VDM
The HPD state is a field of the DP-Status VDO. The DP-Status VDO is
embedded in either the DP-Status VDM or the Attention VDM.

The DP-Status happens before the DP-Config. The pin assignment of the
PD partner hasn't been set up yet. Passing through the HPD to the DP
PHY is too early. Racing it may fail the link-training. Should wait
until the DP-Config is done.

BRANCH=None
BUG=b:247005613
TEST=Connect the device through a Type-C dongle to a monitor and verify
the screen. Try the combinations:
 * Plug C0 monitor -> plug C0 dongle -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> unplug C0 monitor -> plug C0 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> plug C1 dongle -> -> plug C1 monitor -> no screen on C1 monitor
   -> unplug C0 monitor -> replug C1 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C1 dongle -> plug C1 monitor
   -> screen on C1 monitor -> unplug C1 dongle -> plug C0 monitor
   -> screen on C0 monitor

Change-Id: I9387568e2a36f3d44ba2efc78f9fcc2b64bd03fd
Signed-off-by: Wai-Hong Tam <waihong@google.com>