zephyr: test: Make get_panic_data_write mockable
The version of get_panic_data_write() that currently gets compiled in is
controlled by an `#ifdef CONFIG_BOARD_NATIVE_POSIX` block, which means
that the actual implementation used in device builds (over 30 SLOCs) is
unreachable in testing, in favor of a stub that directly returns the
pdata_ptr pointer.

This CL removes the stub native posix version of the function and makes the
real version test_mockable. Then, all test binaries that were relying on
the removed stub definition are supplied with a mocked version of
get_panic_data_write() that behaves the same way the stub did by
returning pdata_ptr directly. This frees a new test binary to test the
actual implementation.

BRANCH=None
BUG=None
TEST=./twister

Signed-off-by: Tristan Honscheid <honscheid@google.com>

Change-Id: I8228acb0147b3b3843d302c272d8715d925fddd5