zephyr emul: TCPC partner: Expose CC to extensions
Implement a callback in the TCPC partner common code for changes to
control registers in the TCPCI TCPC emulator. Specifically, when the
TCPC changes ROLE_CONTROL, record the new CC "voltages" exposed by the
TCPC emulator in the partner emulator. Then call any partner extension
callbacks for control changes.

This change does not cause the partner to react to other changes to
TCPCI control registers.

BUG=b:257341564
TEST=twister -s zephyr/test/drivers/drivers.usbc_ctvpd
BRANCH=none
LOW_COVERAGE_REASON=This is an emulator. Later tests will use it.

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: Id27a156ed5db6d805c439a33a157ae13706d541a