Skyrim: Configure mux resets for ANX7483
The ANX7483 is only turned on in S5 state and higher, and will default
to a high power consumption state.  Flag that this mux resets in G3 to
ensure the USB mux code resets it to None.

BRANCH=None
BUG=b:259863457
TEST=on skyrim, ensure the contents of the control register on anx7483
is the same on first boot as it is after going to G3 and rebooting

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I03cbee24b2849565074a7f856480e48a797db72b