zephyr: usb: enable pi3usb9201 driver and emulator automatically
Enable the PLATFORM_EC_BC12_DETECT_PI3USB9201 and
EMUL_BC12_DETECT_PI3USB9201 based on the devicetree configuration. The
two drivers use the same compatible so the emulator one is set so it can
only be enabled in testing.

This means that the normal one may have to be disabled manually.

BRANCH=none
BUG=none
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Id4438d2af72c46ecf5c47c2d347dfb8bb74a0a57