fan: Refactor and test the most common custom fan_percent_to_rpm.
This is in preparation for reusing this implementation for
{ambassador, genesis, moonbuggy, scout}.

The below-named boards are all the boards that have defined
CONFIG_FAN_RPM_CUSTOM.  They are not changed in this changelist;
this comment just describes what refactoring this changelist
could enable.

{kalista, berknip, chronicler, dewatt, endeavour, ezkinil, fizz} use
this or nearly this implementation of a path-dependent
fan_percent_to_rpm, and can be refactored directly.  The path-dependence
enables hysteresis, typically to avoid oscillations in temperature and
fan speed.

{chronicler} performs boxcar smoothing of the input.

{osiris}, which has two fans, and some other boards with multiple
sensors, use a similar but not exactly the same implementation, and
would need additional abstraction.

BRANCH=refactor-fan-percent-to-rpm
BUG=b:252966838,b:191187610,chromium:1383859
TEST=make -j buildall && make -j runhosttests

Signed-off-by: Richard Yeh <rcy@google.com>
Change-Id: I50ad4d78ac1145f92573a417646c1f57b8945463