fan: Rewrite and test the most common custom fan_percent_to_rpm.
temp_ratio_to_rpm_hysteresis uses a sorted fan_table containing a
mapping of temp_ratio (percent, 0-100) to fan rpm. This clarifies the
relationship of temperature to fan speed, while reducing the number of
static variables. The existing temp_ratio expressing the percentage of
cooling needed (temperature from temp_fan_off to temp_fan_max) suggests
the possibility that a proportional-integral-derivative (PID) feedback
controller might have been considered; but the default implementation
is purely linear (mapping to fan speed from min to max) and this
stepwise one just enables hysteresis --- different speeds when the
board is warming up than when cooling down.

This refactoring is in preparation for reusing this implementation for
{ambassador, genesis, moonbuggy, scout}.

{kalista, berknip, chronicler, dewatt, endeavour, ezkinil, fizz} define
CONFIG_FAN_RPM_CUSTOM and use this or nearly this implementation of a
fan_percent_to_rpm with hysteresis.  The hysteresis attempts to avoid
oscillations in temperature and fan speed.

{chronicler} performs boxcar smoothing of the input.

{osiris}, which has two fans, and {jinlon, redrix, and other boards
with multiple sensors}, use a similar but not exactly the same
implementation, and would need additional abstraction.  See:
https://source.chromium.org/search?q=f:ec%2Fb%20fan_step%20-f:board.c

BRANCH=refactor-fan-percent-to-rpm
BUG=b:252966838,b:191187610,chromium:1383859
TEST=make -j run-fan && make -j buildall && make -j runhosttests

Signed-off-by: Richard Yeh <rcy@google.com>
Change-Id: I50ad4d78ac1145f92573a417646c1f57b8945463