system_safe_mode: Publish EC_HOST_EVENT_PANIC when safe mode starts
Publish EC_HOST_EVENT_PANIC when safe mode starts.

The event is published from a deferred function since publishing events
from an ISR is not compatiable with some systems.

The kernel may use this event to clean up before the system is reset
(e.g. sync the drive).

This will be a no-op if the kernel doesn't handle it.

BUG=b:258195448
BRANCH=None
TEST=Observe event in kernel. Pass panic_event unit test.

Change-Id: I62c48590029bf2bf8f1e0bb7271f29499a8a28d7
Signed-off-by: Rob Barnes <robbarnes@google.com>