system_safe_mode: Publish EC_HOST_EVENT_PANIC when safe mode starts
Publish EC_HOST_EVENT_PANIC when safe mode starts.

The event is published from a deferred function since publishing events
from an ISR is not compatible with some systems.

The kernel may use this event to clean up before the system is reset
(e.g. sync the drive). This will be a no-op if the kernel doesn't handle
it.

This is a refactor of CL:4063818, which was reverted in CL:4225911.
The original implementation triggered the event in the panic handler.
This version triggers the event after safe mode starts. Since the system
reboots immediately when safe mode isn't started, triggering the event
outside of safe mode doesn't have any effect.

BUG=b:258195448
BRANCH=None
TEST=Observe event in kernel. Pass system_safe_mode unit test.

Change-Id: I62c48590029bf2bf8f1e0bb7271f29499a8a28d7
Signed-off-by: Rob Barnes <robbarnes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4226407
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>