TCPMV2: Handle errors in pe_send_alert_run
If the EC tries to send 2 alert messages very close to each other
(~30ms), it will fail with a message discarded error on the second
Alert. Because pe_send_alert_run does not check for this case, it
prevents the TCPMV2 PE from returning to PE_SNK_Ready. This CL updates
pe_send_alert_run to look for protocol errors and discarded messages,
and send a soft reset in either case. This allows the PE to return to
PE_SNK_Ready.

BUG=b:260912784
TEST=Tested by adding EC command to send 2 Alerts very quickly, and
confirmed that if the second alert fails the Chromebook resends it.
Also, "make try_build_boards", "make runhosttests" and zephyr tests.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: I5cc0920a0f910c67e988a3dbf8ac0c3c611c386b