cq: Run steps in parallel by moving logic to make
In order to get everything to run in parallel, just migrate most of the
test logic to a makefile, and call make from the python script.

When building locally, this saved 30% of the time, and got the same
coverage metrics, so I assume the same tests and builds were run.

BRANCH=None
BUG=None
TEST=Ran builder locally, compared metrics with prior commit.

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I7429c35e4a30ad05e092d5aa73d523f3c676e8e7