spi: Allow boards to handle non-standard SPI busses
The STM32L5 series has a special "OctoSPI" controller, which differs
from the standard spi controllers in capabilities and means of
programming.  Rather than adding drivers for OctoSPI to the common EC
code, this change allows special-purpose drivers for HyperDebug to be
implemented in the board directory.

BUG=b:192262089
TEST=Observed HyperDebug forwarding to OctoSPI controller

Change-Id: I317155498ffde77c629bf237810644c8aac4fd45


spi: Forwarding to multiple spi devices

In case of multiple entries in spi_devices, this change allows the USB
host to request which device its transaction requests are to be
forwarded to.

BUG=b:192262089
TEST=HyperDebug board forwards to two SPI busses

Change-Id: I56c7efaf0e82eb7bcc34ae3483b06431b66ef9d4