ap_pwrseq: Do not backup already-cleared SCI and SMI hostevent masks
In some situations, the S0S0ix power state is entered multiple times
ahead of a resume to the point that the kernel sends a host event
indicating the resume, where the SCI and SMI hostevent masks are
restored. When the masks are cleared more than once, the previous
backups are overwritten by the already-cleared values, and the original
masks are lost.

BUG=b:261318939
TEST=hostevent SCI mask is restored after S0ix with a touchpad click
     while the lid is closed
BRANCH=nissa

Change-Id: I8e93ed79a745f7f4c525fb7a8a419198d086448f
Signed-off-by: Sam McNally <sammc@chromium.org>