usb_prl: support logging to a ring buffer
This allows PRL state logging to be displayed asynchronously, with
minimal effect on performance during operation and making it more
tractable to debug timing-sensitive behavior. Events are logged
unconditionally if logging is enabled, and displayed with the `prllog`
console command.

Output looks like this:

5489349 C0 PRL_TX_DISCARD_MESSAGE PRL_TX Set 0x4 : PRL_FLAGS_WAIT_SINK_OK |
5489440 C0 PRL_TX_PHY_LAYER_RESET PRL_TX Set 0x4 : PRL_FLAGS_WAIT_SINK_OK |
5489532 C0 PRL_TX_WAIT_FOR_MESSAGE_REQUEST PRL_TX Set 0x4 : PRL_FLAGS_WAIT_SINK_OK |

Fields are the event timestamp (in microseconds), followed by the port and
state name with the state machine flags on the rest of the line.

BUG=b:229692652
TEST=prllog on nereid console prints correct-looking states
     when CONFIG_PLATFORM_EC_USB_PD_PRL_EVENT_LOG=y
BRANCH=none

Change-Id: I5ba9f0ee65613e5889b606484a62fe60ebe0b2b8
Signed-off-by: Peter Marheine <pmarheine@chromium.org>