zephyr: mtlrvpp_npcx: EC doesn't wakes from hibernate
EC doesn't wakes from hibernate when typec charger is connected on any
of the port. Its noticed that while bc_acok is not toggling we are seeing
the typec interrupts toggle.

Configuring tcpc interrupt as hibernate wake pin works and EC is coming
out of hibernate state. We can use this as workaround until hardware
team rootcause the reason why bc_acok is not toggling

BRANCH=none
BUG=none
TEST=zmake build mtlrvpp_npcx
TEST=EC wakes up from hibernate when typec charger is connected

Change-Id: Ia0a23bdd13e33f455006c0216e02718b004caf22
Signed-off-by: Rajesh Kumar <rajesh3.kumar@intel.com>