zephyr: mtlrvpp_npcx: add TCPC interrupts as hibernate wake pins
EC doesn't wakes from hibernate when Type-C charger is connected on any
of the port.
Its noticed that while bc_acok is not toggling we are seeing the Type-C
interrupts toggle. Configuring tcpc interrupt as hibernate wake pin
works and EC is coming out of hibernate state.

Adding this workaround until actual fix is added in the hardware.

BRANCH=none
BUG=none
TEST=zmake build mtlrvpp_npcx
TEST=EC wakes up from hibernate when Type-C charger is connected

Change-Id: Ia0a23bdd13e33f455006c0216e02718b004caf22
Signed-off-by: Rajesh Kumar <rajesh3.kumar@intel.com>