usb/pe: support logging to a ring buffer
This allows PE state logging to be displayed asynchronously, with
minimal effect on performance during operation and making it more
tractable to debug timing-sensitive behavior. Events are logged
unconditionally if logging is enabled, and displayed with the `pelog`
console command.

Output looks like this:

3389892 C0 2400201 PE_DRS_Change
3391876 C0 2400201 PE_SNK_Ready
3393798 C0 400201 PE_VCS_Send_Swap
3406463 C0 400201 PE_VCS_Force_Vconn
3413848 C0 400201 PE_SNK_Ready
3415802 C0 401201 PE_VDM_Identity_Request_Cbl
3429870 C0 400201 PE_SNK_Ready

Fields are the event timestamp (in microseconds), followed by the port and
port flags (in hex), and finally the PE state being entered.

BUG=b:229692652
TEST=pelog on nereid console prints correct-looking states
     when CONFIG_PLATFORM_EC_USB_PD_PE_EVENT_LOG=y
BRANCH=none

Change-Id: I3665faeb34b1984209f3448a1d9eb3b9a33afbbf
Signed-off-by: Peter Marheine <pmarheine@chromium.org>