zephyr: Provide SLP_S0 DEBUG feature.
We've been using host RTC as an alarm to wake up host for s0ix test.
However, the host RTC alarm won't be canceled when s0ix is failed.
The host RTC wake will break the failure and make the debug more
diffcult.

In order to deep dive s0ix issue, we need the system to stay at failure
and hook up XDP. When system is out of failure, it can be very diffcult
to find out the exact reason blocking s0ix.

here are the cmds to set alarm
1. ectool slps0dbgsetalarm seconds to set alarm
2. ectool slps0dbgsetalarm 0 to cancel alarm

When the alarm is set but SLP_S0 is not asserted before the sleep timeout,
EC will not wake the system, then we can do the SOC debug.

BUG=None
TEST=Tested on screebo and the system can be awaken by EC with time given
     by ectool rtcsetalarm seconds, also can stay at failure if SLP_S0 is
     not assserted on time.

Co-authored-by: Chia-Sheng <chiasheng.lee@intel.corp-partner.google.com>

Change-Id: I3674a81516d1586c3cd0aca83d3dd31a7b836ce9