corsola: bring forward the fake type-c port count init timing
Corsola emulates HDMI MUX over the current type-c MUX stack.
The EC has to fake a type-c port to pass the MUX info.
However, tcpc/pd tasks are not needed for the HDMI port.
We want the tasks are not initiated on starting up, and increase the
type-c port counts after the tasks are launched.

However, on the normal image, the kernel uptime is much faster than on
the test image. There is a chance that the cros-ec-typec getting
the actual type-c port counts from board_get_usb_pd_port_count(),
before we are able to fake it, and this will block the HDMI mux
function.

BUG=b:262898170
TEST=reboot EC for 1900 times and no HDMI output errors
BRANCH=none

Change-Id: I8628c456c55ede796973ecbd6c720e2b2ac7d175
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>