Workaround: winterhold: tune the threshold
The resolution of the sensor lis2dw12 is 14 bits, but the raw data will
be normalized to 16 bits. It cause the body detection variance is always
greater than threshold.

This change reture the 16 bits resolution when body detection enable,
also modify the winterhold CONFIG_PLATFORM_EC_BODY_DETECTION_*
parameters to tune the threshold.

BRANCH=None
BUG=b:262680246
TEST=winterhold can exit body detection mode

Change-Id: I3cc5c0ffa73fbcd318b06e0aa093ec0fd26e5aeb
Signed-off-by: Josh-Tsai <josh_tsai@compal.corp-partner.google.com>