zephyr/test: Wait for watchdog to trigger at the end of test
Every call to watchdog_init() uses one counter channel to run watchdog.
Unfortunately there is no way to disable watchdog running on the channel
to get it back (it's possible to stop counting, but it doesn't free the
channel). Channel can be reused only when watchdog associated with the
channel triggers.

This patch adds waiting for watchdog to trigger (if it hasn't triggered
alread) in watchdog_after() function.

BUG=b:239712345
BRANCH=none
TEST='twister -T zephyr/test/drivers'

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I239e330536502554a7523f7cf0433f594ed233b1