run_device_tests: Add FPMCU power utilization tests
Add tests for checking FPMCU power utilization in both sleep
and idle states. Note that these tests operate on the standard
application image and thus do not require a separate build target
to be compiled. For this reason there is no entry in the
`board/$BOARD/build.mk` and `make BOARD=${BOARD} tests` will not
generate an image for these tests. Running the tests via
`test/run_device_tests.py` will result in the application image
being built, if required.

BUG=b:180445334
TEST=test/run_device_tests.py -b dartmonkey -t power_utilization_sleep
TEST=test/run_device_tests.py -b dartmonkey -t power_utilization_idle
TEST=test/run_device_tests.py -b bloonchipper -t power_utilization_sleep
TEST=test/run_device_tests.py -b bloonchipper -t power_utilization_idle
TEST=test/run_device_tests.py -b helipilot -t power_utilization_sleep
TEST=test/run_device_tests.py -b helipilot -t power_utilization_idle
     NOTE: idle test failing on helipilot, see b/319314358

Change-Id: Ibd17ae1176eaf910c458a87b838c37dcfb24e0aa