TCPMv2: Wake PD task on GoodCRC
When the TCPC alerts to indicate that a message transmitted
successfully, wake the PD task. Shorten the window during which the PRL
can be forced to handle a GoodCRC and the following incoming message in
the same task loop, thereby shortening the window during which the PRL
can check the data role on an incoming message incorrectly following a
Data Role Swap.

BUG=b:262345042
TEST=make buildall
BRANCH=none

Change-Id: Ib59a92752e637b2a8829a90ab1d70755dd324cb9
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>