PRESUBMIT: integrate upstream bindings and yaml check
Add a wrapper script to use the upstream Zephyr check_compliance.py
script in the EC code base.

That script has various presubmit type of checks that are used upstream,
and we can reuse some of those internally as is. The wrapper handles the
vpython dependencies, skips if there's no Zephyr files touched and runs
the YAMLLint and DevicetreeBindings.

BRANCH=none
BUG=b:264661333
TEST=create some test patches to touch the non compliant files, ran the
script on those

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ibe77e92092013780befa54e6c02ab17101b014fb