PRESUBMIT: integrate upstream compliance check - yamllint
Add a wrapper script to use the upstream Zephyr check_compliance.py
script in the EC code base.

That script has various presubmit type of checks that are used upstream,
and we can reuse some of those internally as is. The wrapper handles the
vpython dependencies, skips if there's no Zephyr files touched and only
runs YAMLLint for now.

BRANCH=none
BUG=b:264661333
TEST=create some test patches to touch the non compliant files, ran the
script on those

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Cq-Depend: chromium:4212763
Change-Id: Ibe77e92092013780befa54e6c02ab17101b014fb