run_device_tests: Wrap calls to subprocesses
Replace several instances of duplicated code that prints a command and
then run it via subprocess.run with a function that logs the command and
executes it.

Additionally, utilize a single method of raising a CalledProcessError if
the subprocess has a non-zero return code as all three of the methods
below result in the same behavior:
  * subprocess.run(cmd, check=True)
  * subprocess.run(cmd, check=False).check_returncode()
  * ret = subprocess.run(cmd, check=False)
    ret.check_returncode()

The latter two options were both previously used but the first of
those options seems to be the preferred method.

BUG=none
TEST=run_device_tests.py -b dartmonkey

Change-Id: Ie12c95d54560ac9c07cf57b4b95a6040aa401fd6