usb_mux: If already initialized, do not init muxes
Check the 'initialized' flag for the port, and only initialize
the muxes on the port if they have not already been initialized.

This avoid clearing or resetting the state of the muxes (esp.
when in low power mode) unnecessarily.

If the muxes were reset to their default state after they
were set to low-power-mode, and the AP suspended, the change in their
state (to low power again) sent a host event to the AP, causing the
AP to wake up immediately.

BUG=b:262338992
TEST=Verify on nirwen that AP suspends after hub is unplugged.
BRANCH=none

Change-Id: I15ba40fe978c2f8773b179962de95904bee6b620
Signed-off-by: Andrew McRae <amcrae@google.com>