TEST: Override NVME and EMMC FW_CONFIG bits based on battery_pres pin
craask has an EMMC_SKU_DET strap pin which signals whether or not the
current board has EMMC or NVME storage. In the factory, the CBI images
are prepared with both NVME and EMMC masks set to enable.
Thus this patch will read the FW_CONFIG value at init time,
and if either both NVME and EMMC are selected, or
neither, then it will read the EMMC_SKU_DET pin and set the two
FW_CONFIG masks accordingly.

BUG=b:259211172
BRANCH=none
TEST=check sku workable.

Change-Id: I6467b1fb47f3332e5ac8b560946c43d48348b1b2
Signed-off-by: Ben Chen <ben.chen2@quanta.corp-partner.google.com>