zephyr: fwid: Move RW FWID to end of the flash
Force the RW_FWID to always reside in the last 32 bytes of the flash.
This ensures that the RW_FW and RW_FWID sections never move within the
full EC binary.

The FMAP is only stored in RO, so this ensures future compatibility with
RW firmware image size changes.

BUG=b:264229880
BRANCH=none
TEST=Run "futility dump_fmap ec.bin" and verify RW_FWID is located in
last 32-bytes.
TEST=On Herobrine, validate EC sync and validate FMAP manifest when the
RO and RW images are different versions.

Change-Id: I2d96d602f37ca9fafe4d01cbf1a578e20861abaa
Signed-off-by: Keith Short <keithshort@chromium.org>