panic_output: Allow uart time to send on panic
It was observed that stm32f4 will sometimes fail to transmit the data in
the UART data register on panic_reboot. This was first noticed on the
abort test on the bloonchipper target and it is not clear if other
targets will be similarly affected but it is unlikely. This issue is
believed to be caused by uart_tx_flush looping on STM32_USART_SR.TXE
which only ensures that the transmit data register is empty.
STM32_USART_SR.TC is required to ensure that the transmission was
successfully completed. After a few minutes of updating to utilize TC it
was clear that more time would be needed. As a short term workaround,
provide a short delay prior to  panic_reboot to allow the data to send.

BUG=b:264674276
BRANCH=none
TEST=run_device_tests.py -b bloonchipper -t abort

Change-Id: Ib4a89666612e93fc804346eebf7b3a23fa37d6c1
Signed-off-by: Bobby Casey <bobbycasey@google.com>