chip/stm32: Allow uart time to send on panic
It was observed that stm32f4 will sometimes fail to transmit the data
in the UART data register on panic_reboot. This was observed on the
abort test on the bloonchipper target and is not believed to impact
any other boards. The root cause is believed to be that the debug uart
is flushed via uart_tx_flush looping on STM32_USART_SR.TXE at
shutdown. This ensures that the transmit data register is empty but
does not guarantee that the transmission has successfully
completed, as may be indicated by STM32_USART_SR.TC.

Per the stm32f4 reference manual:
"After writing the last data into the USART_DR register, it is
mandatory to wait for TC=1 before disabling the USART or causing the
microcontroller to enter the low power mode (see Figure 248: TC/TXE
behavior when transmitting)."

BUG=b:264674276
BRANCH=none
TEST=run_device_tests.py -b bloonchipper

Change-Id: Ib4a89666612e93fc804346eebf7b3a23fa37d6c1
Signed-off-by: Bobby Casey <bobbycasey@google.com>