battery: extend BATTERY_GET_STATIC hostcmd for 32-byte strings
The Smart Battery Specification defines a string as up to 31 characters,
and nissa project has encountered some batteries where the existing
11-character strings for v1 of BATTERY_GET_STATIC are insufficient to
differentiate some batteries that can be used on a given device. This
change adds a new version of this host command that extends the strings
fields to each accommodate up to 31 characters of data, which should be
enough to support all smart batteries compliant with the latest version
of the specification.

BUG=b:263920422
TEST=`ectool battery` with updated ectool works on Nereid and uses the
     new hostcmd version. If EC support for v2 or v1 is disabled, older
     command versions are used.
BRANCH=nissa

Change-Id: Ia6a3a917051d5e641e7d38ef6936ec7174cb8aab
Signed-off-by: Peter Marheine <pmarheine@chromium.org>