console: Make it obvious when printk is used for console output
Prepend any console messages printed by printk with a marker
so that it is obvious.

Console output uses printk when invoked from an ISR, and since
printk is unbuffered and no mutual exclusion is used, it can be
expensive.

By highlighting the messages, they can be considered for
converting to Zephyr logs.

BUG=b:265594963
TEST=Verify that printk messages are marked as desired (e.g DPTF).
BRANCH=none

Change-Id: Id743cf1ff46dc2d2da0e5c49a96a07378ca7021f
Signed-off-by: Andrew McRae <amcrae@google.com>