TCPMv2: Snoop AP-driven mode enter
Set up the EC to snoop the AP's mode entry packets for the EnterMode
command, and set up its internal tracking accordingly.  Similarly, watch
the ACK/NAK replies which come so the EC is ready to react to
VDM:Attention messages and send Exit Mode on shutdown.

BRANCH=None
BUG=b:208884535
TEST=on skyrim ensure monitor works with ectool script to
enter and configure DP and EC exits DP mode on shutdown

Change-Id: I1487d3a34c36de4fe9d36556392501edcb9ea9d0
Signed-off-by: Diana Z <dzigterman@chromium.org>