TCPMv2: Rearrange VDM response HC contents
The AP would like to consume VDM:Attention packets in the same host
command as the VDM responses.  To support this, pull the previous return
error into a response structure field instead.

Also correct an incorrect struct for the VDM response test utility
(though it is functionally equivalent to the TYPEC_STATUS struct).

BRANCH=None
BUG=b:208884535
TEST=./twister -T ./zephyr/test

Change-Id: I5d976dda6be40b72de15e84f4bbd9311d908a904
Signed-off-by: Diana Z <dzigterman@chromium.org>