WIP: board/hyperdebug: Support switching into DFU bootloader
Most STM32 chips come with a ROM bootloader, which supports receiving a
firmware image via USB, UART, SPI or I2C.  This bootloader will
activate, if the BOOT0 pin is pulled high as the chip comes out of
reset, which is useful when debugging on the bench.  However, for end
users of HyperDebug, it is desirable to be able to upgrade via USB
without any jumper wires.

This CL introduces a custom console command, which will trigger a jump
from EC firmware into the ROM DFU bootloader.

(The standard does specify a way to trigger the transition via USB
request, we may chose to implement this at a later time, the mechanics
of resetting and jumping would remain the same.)

BUG=none
TEST=Use dfu-util after running new custom command

Change-Id: I2bbcf644294d889c6f1ba7266277058702d538d5