raa489000: Support AC_PRESENT when OTG mode
Flip the comparator inverter output when the charger
chip is running in OTG mode so that the AC_PRESENT
signal is correctly generated. The AC_PRESENT relies on the
ACOK signal, but during OTG mode, ACOK is on when the
charger is sourcing power to the port, so AC_PRESENT should
be off.

BUG=b:263691951
TEST=Verify with nivviks that AC_PRESENT is generated correctly
BRANCH=none

Change-Id: I004b8caaa63a902dd54288840c044bbe920e1347
Signed-off-by: Andrew McRae <amcrae@google.com>