Zephyr: Remove binman.dtsi requirements
Starfish uses the EC build environment but does not
use the CrOSEC stubs and does not use the EC partition.
This leads to linker errors when building due to
CONFIG_PLATFORM_EC_RW_TOTAL_IMAGE_SIZE dependencies
which is undefined on Zephyr only targets.

Adds defined() guards to checks in the linker to avoid
build failures.

BUG=b:243700103
BRANCH=None
TEST=None
Change-Id: I98b207f25edd4f6306fa9ed3a0c1a0ce91f46856
Signed-off-by: Brian Nemec <bnemec@google.com>