common/hooks.h: Tolerate same function registered as multiple hooks
The file clock-stm32l4.c registers the function clock_chipset_startup as
both HOOK_CHIPSET_STARTUP and HOOK_CHIPSET_RESUME.  Before this change,
compiling with hooks disabled would lead to a linker error as a function
unused_hook_clock_chipset_startup() would be declared twice.  This
change makes it such that the hook, as well as the name of the
registered function is concatenated into the synthetic function name.

BUG=b:192262089
TEST=make BOARD=hyperdebug

Change-Id: Ic7ead88e2fd86d509769a1c0a88a788b72576238