chip/stm32: Modification to DFU for STM32L5
The STM32L5 series behaves the same as STM32L4 to a large degree, so
this CL adds code to enable RTC clock on L5 using the same register bits
as on L4.

The "backup registers" however, sit in a new "Tamper" component on L5,
and not in the RTC component as on L4. This CL adds register
declarations for the TAMP component.

Lastly, jumping to the ROM bootloader with C function pointer
call, as was done in jump_to_dfu(), does not work on the STM32L5 ROM
bootloader (at least not on my Nucleo-X552EZ board.)  The flaw seems to
be that an ARM vector table contains a reset vector, and separately
contains an initial value for the stack register.  The previous code
made no attempt at setting the stack point as the bootloader expected,
and depending on how the EC code (and the STM ROM bootloader) is
compiled, this may or may not work.  This CL adds assembly code to set
the stack pointer to the value given in the ROM header, before
performing the jump.

BUG=b:192262089
TEST=Upgrade firmware on STM32L5 based Nucleo board using dfu-util

Change-Id: I2a871f0eaf9ea9f4c27c75ccb472cd4ff9616524