util: check_zephyr_project_config: ignore west modules
Currently parse_modules load all Zephyr modules if west is installed on
the system, which breaks this presubmit check if it is run outside of
the chroot and west is installed. Fix that by passing EC_BASE as modules
instead of extra_modules, which prevents the west list to be parsed in
the first place.

BRANCH=none
BUG=none
TEST=./util/check_zephyr_project_config.py -d $( find -name '*.conf' )

Change-Id: I43ec7dff6c2b1a4e289d7afe4a9c807a2646f531
Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>