Starfish: Button Interface and GPIO
Implements a debounced button interface for Starfish, 3 buttons
are present to control the state with short and long button
press events. Hysteresis is present to prevent rapid events
from being generated and once the buttons have been released
we will process the events.

Adds GPIO_LIST_CTX which is a helper macro which generates
a struct mapping to each GPIO associated with the group.
This enables automatic assignment of GPIO labels, indexes,
and creation of state-machines using the device tree.

BUG=b:265980723

BRANCH=None

TEST=Build, flash, and connect the console.
TEST=Validate buttons work and console events are generated.

Change-Id: Id1ba423468dff03ce72260c5d4b9e811f4f6a440
Signed-off-by: Brian Nemec <bnemec@chromium.org>