TCPMv2: Trigger ErrorRecovery for FRS corner cases
Currently, if the port partner sends a Hard Reset during the FRS
sequence, we attempt to handle the Hard Reset.  However, this leaves us
in an inconsistent power state which can lead to the DUT sourcing itself
while neither partner presents Rp.  Instead, go to ErrorRecovery which
is the general method of recovering from anything unexpected during the
FRS sequence.

Additionally, correct the state machine to go to ErrorRecovery on
discard as well and when we receive a SoftReset in the middle of the FRS
sequence.

BRANCH=None
BUG=b:266035645
TEST=on frostflow, trigger FRS with problematic dock and observe Chrome
does not show charging
LOW_COVERAGE_REASON=FRS support in the partner emulator needs expanded,
filed b/269473786 to track this

Change-Id: Icd3a9da52534ecbb9012b740b082938f66052c22
Signed-off-by: Diana Z <dzigterman@chromium.org>