TCPM: Divorce v1 and v2 alternate mode handling
Retire the common DFP mode support to the TCPMv1 specific module it was
once copied from.  Since TCPMv1 does not support TBT, all TBT functions
can be removed from its copy.

Move the DFP support for TCPMv2 to the DPM and its modules.  Note that
the reduced redundancy in tracking active alternate modes means that we
can remove the majority of the "generic" mode handling.  This results in
approximately 1k additional free flash space for boards.

For now, leave the svdm_dp_* interfaces in place as boards are using
them as entry points to modify the DP mode sequencing.  Also leave a
stripped down version of pd_dfp_exit_mode() in place, though this would
be worth consolidating to a new API in the future.

BRANCH=None
BUG=b:170372521,b:159856063
TEST=all unit tests passing, confirm display works on nipperkin (ECOS)
and skyrim (zephyr) boards

Change-Id: Idbe59bc7c6a0ab6103e8fa158e69275a142f8a16
Signed-off-by: Diana Z <dzigterman@chromium.org>